Setting queue filter on plugin initialization
Ensuring the realtime queue stats view does not subscribe to all task queues by default, which causes performance issues when there are 100+ queues. Also now pulling the queue filter TaskQueue list from the TaskQueue API so it's not dependent on queue stats subscribing to all queues to populate the list.